Refactored message structs to use int instead of size_t

+++ GTmyMusic.h
@@ -39,25 +39,25 @@
 
 	typedef struct {
 		char command;
-		size_t filenames_length;
+		int filenames_length;
 		char* serialized_server_filenames;
 	} diff_message;
 
 	typedef struct {
 		char command;
-		size_t filenames_length;
+		int filenames_length;
 		char* serialized_server_filenames;
 	} pull_message_1;
 
 	typedef struct {
 		char command;
-		size_t filenames_length;
+		int filenames_length;
 		char* serialized_client_filenames;
 	} pull_message_2;
 
 	typedef struct {
 		char command;
-		size_t files_length;
+		int files_length;
 		FILE* server_files;
 	} pull_message_3;
